code/neverending/src/app/requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.0.0
databricks-sdk>=0.35.0
databricks-sql-connector>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
